backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
celery>=5.3.0
redis>=4.5.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0

# API Documentation
drf-spectacular>=0.27.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
